pedalpal/requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==4.2.10
django-adminlte-3==0.1.6
django-rest-passwordreset==1.4.0
django_rest_resetpassword==0.1.2
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.34
nodeenv==1.8.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.1
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.1
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.25.0
asgiref==3.7.2
cfgv==3.4.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.2.10
django-adminlte-3==0.1.6
django-import-export==3.3.7
django-phonenumber-field==7.3.0
django-rest-passwordreset==1.4.0
django_rest_resetpassword==0.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
identify==2.5.34
MarkupPy==1.14
nodeenv==1.8.0
odfpy==1.4.1
openpyxl==3.1.2
phonenumberslite==8.13.30
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.1
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
virtualenv==20.25.0
xlrd==2.0.1
xlwt==1.3.0
